* debian: (29 commits)
  30_wireless_support.diff
  29_syspower_battery_support.diff
  28_sh4_support.diff
  27_fix_const.diff
  26_spelling_fix.diff
  25_iostream_h_FTBFS_fixes.diff
  24_enhanced_memory.diff
  23_cpumeter_guest_time.diff
  22_fix_valgrind_reports.diff
  21_fix_unclutter.diff
  20_fix_swapmeter.diff
  19_fix_reading_interrupts.diff
  18_handle_high_irq_numbers.diff
  17_cpuFormat-linux-support.diff
  16_lmstemp2.6.diff
  15_FTBFS_alpha_missing_irq.diff
  14_gnu_hurd_support_436136.diff
  13_manpage_fixes.diff
  12_stolen_time_support.diff
  11_float_to_double_151611.diff
  ...

Conflicts:
	Xdefaults.in
	Xdefaults.stipple
	Xrm.cc
	config/aclocal.m4
	config/configure.in
	gnu/MeterMaker.h
	gnu/loadmeter.h
	gnu/memmeter.h
	gnu/swapmeter.h
	linux/MeterMaker.cc
	linux/cpumeter.cc
	linux/cpumeter.h
	linux/serialmeter.cc
	xwin.cc

+++ linux/swapmeter.cc
@@ -54,7 +54,7 @@ void SwapMeter::checkevent( void ){
 #ifdef USESYSCALLS
 void SwapMeter::getswapinfo( void ){
   struct sysinfo sinfo;
-  int unit;
+  unsigned long unit;
 
 #if defined(GNULIBC) || defined(__GLIBC__)
   sysinfo(&sinfo);
@@ -64,9 +64,9 @@ void SwapMeter::getswapinfo( void ){
 
   unit = (sinfo.mem_unit ? sinfo.mem_unit : 1);
 
-  total_ = sinfo.totalswap * unit;
-  fields_[0] = (sinfo.totalswap - sinfo.freeswap) * unit;
-  fields_[1] = sinfo.freeswap * unit;
+  total_ = (double)sinfo.totalswap * unit;
+  fields_[0] = (double)(sinfo.totalswap - sinfo.freeswap) * unit;
+  fields_[1] = (double)sinfo.freeswap * unit;
 
   if ( total_ == 0 ){
     total_ = 1;
